+++ src/main/java/kitchenpos/application/ProductService.java
@@ -1,8 +1,8 @@
 package kitchenpos.application;
 
 import java.util.List;
-import kitchenpos.domain.product.ProductRepository;
-import kitchenpos.domain.product.Product;
+import kitchenpos.domain.menu.Product;
+import kitchenpos.domain.menu.ProductRepository;
 import kitchenpos.dto.request.ProductRequest;
 import kitchenpos.dto.response.ProductResponse;
 import org.springframework.stereotype.Service;
