+++ src/main/java/kitchenpos/application/TableService.java
@@ -3,9 +3,9 @@ package kitchenpos.application;
 import java.util.Arrays;
 import java.util.List;
 import kitchenpos.domain.order.OrderRepository;
-import kitchenpos.domain.ordertable.OrderTableRepository;
 import kitchenpos.domain.order.OrderStatus;
 import kitchenpos.domain.ordertable.OrderTable;
+import kitchenpos.domain.ordertable.OrderTableRepository;
 import kitchenpos.dto.request.OrderTableRequest;
 import kitchenpos.dto.response.OrderTableResponse;
 import org.springframework.stereotype.Service;
