+++ src/main/java/kitchenpos/domain/menu/Menu.java
@@ -7,7 +7,6 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import kitchenpos.domain.Price;
 
 @Entity
 public class Menu {
