+++ src/main/java/kitchenpos/domain/menu/Product.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain.product;
+package kitchenpos.domain.menu;
 
 import java.math.BigDecimal;
 import javax.persistence.Embedded;
@@ -6,7 +6,6 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import kitchenpos.domain.Price;
 
 @Entity
 public class Product {
