+++ src/test/java/kitchenpos/acceptance/MenuAcceptanceTest.java
@@ -1,11 +1,11 @@
 package kitchenpos.acceptance;
 
 import io.restassured.response.ValidatableResponse;
-import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
 import kitchenpos.domain.menu.MenuGroup;
 import kitchenpos.domain.menu.MenuProduct;
-import kitchenpos.domain.product.Product;
+import kitchenpos.domain.menu.Product;
 import kitchenpos.dto.request.MenuRequest;
 import kitchenpos.support.RequestBuilder;
 import org.hamcrest.Matchers;
@@ -23,7 +23,7 @@ public class MenuAcceptanceTest extends AcceptanceTest {
         final int price = 4000;
         final Product savedProduct = dataSupport.saveProduct("참치마요", price);
         final MenuGroup savedMenuGroup = dataSupport.saveMenuGroup("할인 상품");
-        final List<MenuProduct> menuProducts = Arrays.asList(MenuProduct.ofUnsaved(null, savedProduct, 1L));
+        final List<MenuProduct> menuProducts = Collections.singletonList(MenuProduct.ofUnsaved(null, savedProduct, 1L));
         final int discountedPrice = price - 500;
 
         // when
