+++ src/test/java/kitchenpos/domain/menu/MenuTest.java
@@ -3,7 +3,6 @@ package kitchenpos.domain.menu;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.math.BigDecimal;
-import kitchenpos.domain.Price;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
